remove inline styles, isolate python functions and make edits according to review
