+++ static/css/legacy.less
@@ -37,6 +37,7 @@
 @import (less) "components/header.less";
 @import (less) "components/mybooks.less";
 @import (less) "components/readinglog-stats.less";
+@import (less) "components/tagging-menu.less";
 @import (less) "components/shareLinks.less";
 @import (less) "components/dropper.less";
